+++ examples/tauth-demo/demo-config.js
@@ -1,0 +1,8 @@
+// @ts-check
+'use strict';
+
+const DEMO_CONFIG = Object.freeze({
+  googleClientId: '',
+});
+
+window.__TAUTH_DEMO_CONFIG = DEMO_CONFIG;